requirements.txt
# pywin32 — Clipboard (win32clipboard / win32con)
pywin32>=306
# tkinter — Standardbibliothek unter Windows (kein pip-Paket)

# Anker-Kalibrierung (--calibrate-anchor): PIL Image / ImageTk für Screenshot-Dialog
Pillow>=10.0.0
# Vermessung: OpenCV Template-Matching + Kanten-Heuristik für Textboxen
opencv-python-headless>=4.8.0
numpy>=1.24.0
